feat: refine infrastructure dashboard
