+++ src/app/(main)/dashboard/infrastructure/page.tsx
@@ -1,6 +1,9 @@
 import { infrastructureGroups } from "./_components/infrastructure-data";
 import { InfrastructureHeader } from "./_components/infrastructure-header";
-import { InfrastructureProjectGroup } from "./_components/infrastructure-project-group";
+import { ProjectEnvironments } from "./_components/project-environments";
+
+// Import this stylesheet in any page or component that renders country flag classes.
+import "@/styles/flag-icons/flags.css";
 
 export default function Page() {
   return (
@@ -9,7 +12,7 @@ export default function Page() {
 
       <div className="flex flex-col gap-4">
         {infrastructureGroups.map((group) => (
-          <InfrastructureProjectGroup key={group.name} group={group} />
+          <ProjectEnvironments key={group.name} group={group} />
         ))}
       </div>
     </div>
